team_40/requirements.txt
setuptools
wheel
click==8.1.7
Flask==2.3.3
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy
pandas
Pillow
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
typing-extensions==4.8.0
Werkzeug==2.3.7
torch==2.2.0
torchvision==0.17.0
python-dotenv      # <-- add this line

Flask-PyMongo
flask-login
Flask-Mail
Flask-SocketIO
